rearrange: create cp module
